+++ typescript/ui/src/core/contexts/ConfigContext/ConfigContextProvider.tsx
@@ -1,13 +1,11 @@
 import { PropsWithChildren, useEffect, useMemo } from "react";
 import { ConfigContext } from "./ConfigContext";
 import { useRagbitsCall } from "@ragbits/api-client-react";
-import { useThemeContext } from "../ThemeContext/useThemeContext";
 import { CircularProgress, cn } from "@heroui/react";
 import { FeedbackFormPluginName } from "../../../plugins/FeedbackPlugin";
 import { pluginManager } from "../../utils/plugins/PluginManager";
 
 export function ConfigContextProvider({ children }: PropsWithChildren) {
-  const { theme } = useThemeContext();
   const { call: fetchConfig, ...config } = useRagbitsCall("/api/config");
 
   const value = useMemo(() => {
@@ -42,7 +40,6 @@ export function ConfigContextProvider({ children }: PropsWithChildren) {
       <div
         className={cn(
           "flex h-screen w-screen items-start justify-center bg-background",
-          theme,
         )}
       >
         <div className="m-auto flex flex-col items-center gap-4 text-default-900">
@@ -58,7 +55,6 @@ export function ConfigContextProvider({ children }: PropsWithChildren) {
       <div
         className={cn(
           "flex h-screen w-screen items-start justify-center bg-background",
-          theme,
         )}
       >
         <div className="m-auto flex flex-col items-center gap-4 text-default-900">