+++ typescript/ui/src/plugins/ExamplePlugin/index.tsx
@@ -5,7 +5,7 @@ export const ExamplePluginName = "ExamplePlugin";
 export const ExamplePlugin = createPlugin({
   name: ExamplePluginName,
   components: {
-    ExampleComponent: lazy(() => import("./ExamplePluginComponent")),
+    ExampleComponent: lazy(() => import("./components/ExamplePluginComponent")),
   },
   onActivate: () => {
     console.log("ExamplePlugin activated");